Move markdown to html translation to view models
This is more in line with the handling of raw html for SVG files and
simplifies the views.
