--- src/main/kotlin/nl/avisi/structurizr/site/generatr/site/model/MarkdownViewModel.kt
@@ -1,3 +1,0 @@
-package nl.avisi.structurizr.site.generatr.site.model
-
-data class MarkdownViewModel(val markdown: String, val svgFactory: (key: String, url: String) -> String?)
