+++ src/main/kotlin/nl/avisi/structurizr/site/generatr/site/views/HomePage.kt
@@ -6,7 +6,7 @@ import nl.avisi.structurizr.site.generatr.site.model.HomePageViewModel
 fun HTML.homePage(viewModel: HomePageViewModel) {
     page(viewModel) {
         contentDiv {
-            markdown(viewModel, viewModel.content)
+            rawHtml(viewModel.content)
         }
     }
 }
