+++ src/main/kotlin/nl/avisi/structurizr/site/generatr/site/views/RawHtml.kt
@@ -1,0 +1,13 @@
+package nl.avisi.structurizr.site.generatr.site.views
+
+import kotlinx.html.FlowContent
+import kotlinx.html.div
+import kotlinx.html.unsafe
+
+fun FlowContent.rawHtml(html: String) {
+    div {
+        unsafe {
+            +html
+        }
+    }
+}
