+++ src/test/kotlin/nl/avisi/structurizr/site/generatr/site/model/SoftwareSystemSectionPageViewModelTest.kt
@@ -33,6 +33,6 @@ class SoftwareSystemSectionPageViewModelTest : ViewModelTest() {
         val section = createSection()
         val viewModel = SoftwareSystemSectionPageViewModel(generatorContext, softwareSystem, section)
 
-        assertThat(viewModel.content).isEqualTo(MarkdownViewModel(section.content, svgFactory))
+        assertThat(viewModel.content).isEqualTo(markdownToHtml(viewModel, section.content, svgFactory))
     }
 }
